+++ dbinserter/package.json
@@ -2,9 +2,9 @@
   "name": "dbinserter",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "uploader.js",
   "scripts": {
-    "start": "nodemon index.js",
+    "start": "nodemon uploader.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
