--- dbinserter/test.js
@@ -1,26 +1,0 @@
-const { Entity, Column, PrimaryGeneratedColumn } = require("typeorm");
-
-Entity()
-class Photo {
-    PrimaryGeneratedColumn()
-    id: number
-
-    Column({
-        length: 100,
-    })
-    name: string
-
-    Column("text")
-    description: string
-
-    Column()
-    filename: string
-
-    Column("double")
-    views: number
-
-    Column()
-    isPublished: boolean
-}
-
-module.exports = Photo;
